moved audio implementation into new audio subdir
