+++ src/engine/CMakeLists.txt
@@ -38,6 +38,7 @@ file(
         script/duktape-2.2.0/*.c
         script/duktape-2.2.0/*.h
 
+        audio/audio.cpp
         graphics/*.*
         script/*.*
         game/*.*
