+++ src/engine/game/SPGameMode.cpp
@@ -19,7 +19,7 @@
 
 #include "../engine.h"
 #include "../frontend.h"
-#include "../audio.h"
+#include "../audio/audio.h"
 
 #include "SPGameMode.h"
 #include "ActorManager.h"
